move cleanable and exportable to package used
