+++ src/main/java/io/mattw/youtube/commentsuite/db/YouTubeComment.java
@@ -4,7 +4,6 @@ import com.google.api.services.youtube.model.Comment;
 import com.google.api.services.youtube.model.CommentSnippet;
 import com.google.api.services.youtube.model.CommentThread;
 import com.google.api.services.youtube.model.CommentThreadSnippet;
-import io.mattw.youtube.commentsuite.Exportable;
 import io.mattw.youtube.commentsuite.FXMLSuite;
 import io.mattw.youtube.commentsuite.util.DateUtils;
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
