+++ src/main/java/io/mattw/youtube/commentsuite/fxml/Cleanable.java
@@ -1,4 +1,4 @@
-package io.mattw.youtube.commentsuite;
+package io.mattw.youtube.commentsuite.fxml;
 
 /**
  * For use in instances where UI components can "clean up" and reset their state instead of creating an
