+++ src/main/java/io/mattw/youtube/commentsuite/fxml/SYAddToGroupModal.java
@@ -1,7 +1,6 @@
 package io.mattw.youtube.commentsuite.fxml;
 
 import com.google.common.eventbus.Subscribe;
-import io.mattw.youtube.commentsuite.Cleanable;
 import io.mattw.youtube.commentsuite.FXMLSuite;
 import io.mattw.youtube.commentsuite.db.CommentDatabase;
 import io.mattw.youtube.commentsuite.db.Group;
